Adds --game ("-g") and --asset-dir ("-a").
Add some command line options which can be set in IDE run configurations to
reduce the pain of moving back and forth between Windows and Linux, and to
encourage writing of more demos and minigames.

Being able to change the game allows different demos to be built and run
separately.  Changing the asset dir allows for experimentation of assets or
being able to use of multiple locally stored groups of assets.