Adds CMake and stamp_file.c docs to describe why it's needed.
Knell links to knell-lib, but relies on the build copying knell-lib
into it's directory in a post-build step.  When the knell driver doesn't
need to be rebuilt the post-build step will not fire.  This forces knell
to depend on something which triggers a post-build -- a 0 byte generated
file it #include's.