Rename "Hammer" -> "Crank".
I found out "Hammer" is the name of a tool Valve built for the Source
engine.  It's such a cool name someone is already using it.  Keeping with
the "bell" theme of "knell", manual bells sometimes have a "crank" you use
to sound them, so this tool naming makes sense.